- change data type

+++ src/TheCoach.Web.Models/Binding/PurchaseLessonBindingModel.cs
@@ -4,7 +4,7 @@ namespace TheCoach.Web.Models.Binding
 {
 	public class PurchaseLessonBindingModel
 	{
-		public int LessonId { get; set; }
+		public long LessonId { get; set; }
 		public string StoreProductId { get; set; }
 		public string PurchaseToken { get; set; }
 		public PurchaseState State { get; set; }
